agents/requirements.txt
# Kafka and Schema Registry
kafka-python==2.0.2
fastavro==1.9.0
requests==2.31.0

# MCP (Model Context Protocol) support
mcp==1.0.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Web framework
fastapi==0.95.2
uvicorn==0.22.0

# Workaround for pydantic conflict
pydantic==1.10.12
